Added Exceptions handling

+++ EasySave-V1/Model/Singleton/SaveTemplateConfig.cs
@@ -1,4 +1,5 @@
 using Newtonsoft.Json;
+using System;
 using System.Collections.Generic;
 using System.IO;
 
@@ -39,28 +40,35 @@ namespace NSModel.Singleton {
 		/* Method to add a SaveTemplate in the config file */
 		public void Write(SaveTemplate template) {
 
-			/* Unhiding file to allow edit */
-			var attributes = File.GetAttributes(file.ToString());
-			attributes &= ~FileAttributes.Hidden;
-			File.SetAttributes(file.ToString(), attributes);
+			if(new DirectoryInfo(template.srcDirectory).Exists)
+            {
+				/* Unhiding file to allow edit */
+				var attributes = File.GetAttributes(file.ToString());
+				attributes &= ~FileAttributes.Hidden;
+				File.SetAttributes(file.ToString(), attributes);
 
-			string reader = File.ReadAllText(file.ToString());
+				string reader = File.ReadAllText(file.ToString());
 
-			/* Reading the config file and converting it to a list of SaveTemplates */
-			List<SaveTemplate> templates = JsonConvert.DeserializeObject<List<SaveTemplate>>(reader);
-			if (templates == null)
-			{
-				templates = new List<SaveTemplate>();
+				/* Reading the config file and converting it to a list of SaveTemplates */
+				List<SaveTemplate> templates = JsonConvert.DeserializeObject<List<SaveTemplate>>(reader);
+				if (templates == null)
+				{
+					templates = new List<SaveTemplate>();
+				}
+				templates.Add(template);
+				StreamWriter writer = new StreamWriter(file.ToString());
+
+				/* Converting the list of SaveTemplates to a Json and writing it to config file */
+				writer.WriteLine(JsonConvert.SerializeObject(templates, Formatting.Indented));
+				writer.Close();
+
+				/* Hiding file */
+				File.SetAttributes(file.ToString(), File.GetAttributes(file.ToString()) | FileAttributes.Hidden);
 			}
-			templates.Add(template);
-			StreamWriter writer = new StreamWriter(file.ToString());
-
-			/* Converting the list of SaveTemplates to a Json and writing it to config file */
-			writer.WriteLine(JsonConvert.SerializeObject(templates, Formatting.Indented));
-			writer.Close();
-
-			/* Hiding file */
-			File.SetAttributes(file.ToString(), File.GetAttributes(file.ToString()) | FileAttributes.Hidden);
+			else
+            {
+				throw new Exception("The source directory doesn't exist");
+            }
 		}
 
 		/* Method to delete a SaveTemplate in the config file */
